added method middleware, added all layers for merch, added transactions to repositories

+++ internal/app/models/merch.go
@@ -5,5 +5,5 @@ import "gorm.io/gorm"
 type Merch struct {
 	gorm.Model
 	Name  string
-	Price int64
+	Price int
 }
